requirements.txt
streamlit>=1.33
pandas>=2.0
numpy>=1.24
scipy>=1.10
pyyaml>=6.0
